refactor: theme use useSyncExternal hook
